more changes for test file and function namings
